fix: rename uuid to sid
