+++ lib/twilio/sid.ts
@@ -4,11 +4,12 @@
  * @internalapi
  */
 
-import { NotSupportedError } from '../twilio/errors';
+import { NotSupportedError } from './errors';
 
 /**
- * Generates a 16 character long random string where each character is from the
- * set [0,1,2,3,4,5,6,7,8,9,a,b,c,d,e,f].
+ * Generates a 128-bit long random string that is formatted as a 32 long string
+ * of characters where each character is from the set:
+ * [0,1,2,3,4,5,6,7,8,9,a,b,c,d,e,f].
  */
 function generateRandomizedString(): string {
   if (typeof window !== 'object') {
